bfabric_app_runner/examples/file_table/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "file_table_app"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "pyyaml",
    "snakemake",
    "polars",
    "bfabric",
    "bfabric-app-runner",
    "cyclopts",
    "loguru",
    "pyjanitor",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-mock",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "src/a375_mzmine/**/*.py",
    "src/a375_mzmine/data/**/*.qmd",
    "src/a375_mzmine/data/**/*.yml",
    "src/a375_mzmine/data/**/*.mako",
]
